bugfix/FT-23new: add local storage for forms, autoclear, autocomplete
