+++ src/js/firebaseAuth.js
@@ -6,17 +6,33 @@ import {
 } from './register-modal';
 import { FireBaseApi } from './fireBaseApi';
 
-export async function authSignUpUser(event) {
+const refs = {
+  formSignIn: document.querySelector('#form-sign-in'),
+  formSignUp: document.querySelector('#form-sign-up'),
+};
+
+let userName;
+const FORM_STORAGE_KEY = 'register-form-state';
+
+let formData = {};
+
+refs.formSignIn.addEventListener('submit', authSignInUser);
+refs.formSignUp.addEventListener('submit', authSignUpUser);
+
+refs.formSignIn.addEventListener('input', addFormFields);
+refs.formSignUp.addEventListener('input', addFormFields);
+
+async function authSignUpUser(event) {
   event.preventDefault();
 
-  // if (!formData['user-mail'] || !formData['user-password']) {
-  //   return alert('Please, fill all form fields');
-  // }
+  if (!formData['mail'] || !formData['password']) {
+    return alert('Please, fill all form fields');
+  }
 
-  // localStorage.removeItem(FORM_STORAGE_KEY);
-  // console.log(formData);
-  // event.currentTarget.reset();
-  // clearFormData(formData);
+  localStorage.removeItem(FORM_STORAGE_KEY);
+  console.log(formData);
+  event.currentTarget.reset();
+  clearFormData(formData);
 
   userName = await event.target.querySelector('#user-name').value;
   const email = await event.target.querySelector('#user-email-register').value;
@@ -41,14 +57,14 @@ export async function authSignUpUser(event) {
 export async function authSignInUser(event) {
   event.preventDefault();
 
-  // if (!formData['user-mail'] || !formData['user-password']) {
-  //   return alert('Please, fill all form fields');
-  // }
+  if (!formData['mail'] || !formData['password']) {
+    return alert('Please, fill all form fields');
+  }
 
-  // localStorage.removeItem(FORM_STORAGE_KEY);
-  // console.log(formData);
-  // event.currentTarget.reset();
-  // clearFormData(formData);
+  localStorage.removeItem(FORM_STORAGE_KEY);
+  console.log(formData);
+  event.currentTarget.reset();
+  clearFormData(formData);
 
   const email = event.target.querySelector('#user-email').value;
   const password = event.target.querySelector('#user-password').value;
@@ -66,11 +82,6 @@ export async function authSignInUser(event) {
       }
     });
 }
-
-// create modal form
-// function getAuthFormHTML() {
-//   return ``;
-// }
 
 export async function authWithEmailAndPassword(email, password) {
   const apiKey = 'AIzaSyAysL6D0-x9xs_8XCW-NbLtbtHl5P6b3V0';
@@ -110,20 +121,46 @@ export async function registerUser(email, password, name) {
   );
 }
 
-// function fetch(token) {
-//   if (token) {
-//     return Promise.resolve('Erorr!');
-//   }
-//   return console.log('get info');
-//   //return fetch(link/${token}).then(response => response.json).
-//   // then(list of info => console.log(lsit of info))
-// }
+function clearFormData(obj) {
+  for (const key in obj) {
+    delete obj[key];
+  }
+}
 
-// Authorization
-// {
-// 'rules': { '.read: 'auth != null', '.write':true}
-// }
+function addFormFields(event) {
+  formData[event.target.name] = event.target.value;
+  localStorage.setItem(FORM_STORAGE_KEY, JSON.stringify(formData));
+}
 
-// function renderModalAfterAuth(content) {
-//   console.log(content);
-// }
+function getStorageDataSignIn() {
+  formData = JSON.parse(localStorage.getItem(FORM_STORAGE_KEY));
+  if (formData) {
+    setFormFieldsSignIn(formData);
+  } else {
+    formData = {};
+  }
+}
+
+function getStorageDataSignUp() {
+  formData = JSON.parse(localStorage.getItem(FORM_STORAGE_KEY));
+  if (formData) {
+    setFormFieldsSignUp(formData);
+  } else {
+    formData = {};
+  }
+}
+
+function setFormFieldsSignIn(obj) {
+  for (const key in obj) {
+    refs.formSignIn[key].value = obj[key];
+  }
+}
+
+function setFormFieldsSignUp(obj) {
+  for (const key in obj) {
+    refs.formSignUp[key].value = obj[key];
+  }
+}
+
+window.addEventListener('load', getStorageDataSignIn);
+window.addEventListener('load', getStorageDataSignUp);
