+++ src/js/register-modal.js
@@ -1,5 +1,4 @@
 import { authSignInUser, authSignUpUser } from './firebaseAuth';
-import { FireBaseApi } from './fireBaseApi';
 
 const refs = {
   formSignIn: document.querySelector('#form-sign-in'),
@@ -13,16 +12,6 @@ const refs = {
   buttonChangeModal: null,
   buttonCloseGreetings: null,
 };
-let userName;
-const FORM_STORAGE_KEY = 'register-form-state';
-
-let formData;
-
-refs.formSignIn.addEventListener('submit', authSignInUser);
-refs.formSignUp.addEventListener('submit', authSignUpUser);
-
-refs.formSignIn.addEventListener('input', addFormFields);
-refs.formSignUp.addEventListener('input', addFormFields);
 
 refs.buttonClose = document.querySelector('[data-modal-close-sign-in]');
 refs.buttonClose.addEventListener('click', openSignInModal);
@@ -55,91 +44,63 @@ function changeModal() {
   openSignInModal();
 }
 
-function onFormSignInSubmit(event) {
-  event.preventDefault();
+// function onFormSignInSubmit(event) {
+//   event.preventDefault();
 
-  if (!formData['user-mail'] || !formData['user-password']) {
-    return alert('Please, fill all form fields');
-  }
+//   if (!formData['user-mail'] || !formData['user-password']) {
+//     return alert('Please, fill all form fields');
+//   }
 
-  localStorage.removeItem(FORM_STORAGE_KEY);
-  console.log(formData);
-  event.currentTarget.reset();
-  clearFormData(formData);
+//   localStorage.removeItem(FORM_STORAGE_KEY);
+//   console.log(formData);
+//   event.currentTarget.reset();
+//   clearFormData(formData);
 
-  const email = event.target.querySelector('#user-email').value;
-  const password = event.target.querySelector('#user-password').value;
+//   const email = event.target.querySelector('#user-email').value;
+//   const password = event.target.querySelector('#user-password').value;
 
-  authWithEmailAndPassword(email, password)
-    .then(response => response.json())
-    .then(data => {
-      if (data.error) {
-        alert(data.error.message);
-      } else {
-        console.log('It*s ok!');
-        FireBaseApi.authSuccess(data);
-        console.log(data);
-        openSignInModal();
-      }
-    });
-}
+//   authWithEmailAndPassword(email, password)
+//     .then(response => response.json())
+//     .then(data => {
+//       if (data.error) {
+//         alert(data.error.message);
+//       } else {
+//         console.log('It*s ok!');
+//         FireBaseApi.authSuccess(data);
+//         console.log(data);
+//         openSignInModal();
+//       }
+//     });
+// }
 
-async function onFormSignUpSubmit(event) {
-  event.preventDefault();
+// async function onFormSignUpSubmit(event) {
+//   event.preventDefault();
 
-  if (!formData['user-mail'] || !formData['user-password']) {
-    return alert('Please, fill all form fields');
-  }
+//   if (!formData['user-mail'] || !formData['user-password']) {
+//     return alert('Please, fill all form fields');
+//   }
 
-  localStorage.removeItem(FORM_STORAGE_KEY);
-  console.log(formData);
-  event.currentTarget.reset();
-  clearFormData(formData);
+//   localStorage.removeItem(FORM_STORAGE_KEY);
+//   console.log(formData);
+//   event.currentTarget.reset();
+//   clearFormData(formData);
 
-  userName = await event.target.querySelector('#user-name').value;
-  const email = await event.target.querySelector('#user-email-register').value;
-  const password = await event.target.querySelector('#user-password-register')
-    .value;
+//   userName = await event.target.querySelector('#user-name').value;
+//   const email = await event.target.querySelector('#user-email-register').value;
+//   const password = await event.target.querySelector('#user-password-register')
+//     .value;
 
-  await registerUser(email, password, userName)
-    .then(response => response.json())
-    .then(data => {
-      if (data.error) {
-        alert(data.error.message);
-      } else {
-        console.log('It*s ok!');
-        // FireBaseApi.authSuccess(userName);
-        FireBaseApi.authSuccess(data);
-        openGreetingsModal();
-        openSignUpModal();
-      }
-    });
-}
-
-function clearFormData(obj) {
-  for (const key in obj) {
-    delete obj[key];
-  }
-}
-
-function addFormFields(event) {
-  formData[event.target.name] = event.target.value;
-  localStorage.setItem(FORM_STORAGE_KEY, JSON.stringify(formData));
-}
-
-function getStorageData() {
-  formData = JSON.parse(localStorage.getItem(FORM_STORAGE_KEY));
-  if (formData) {
-    setFormFields(formData);
-  } else {
-    formData = {};
-  }
-}
-
-function setFormFields(obj) {
-  for (const key in obj) {
-    refs.form[key].value = obj[key];
-  }
-}
-
-window.addEventListener('load', getStorageData);
+//   await registerUser(email, password, userName)
+//     .then(response => response.json())
+//     .then(data => {
+//       if (data.error) {
+//         alert(data.error.message);
+//       } else {
+//         console.log('It*s ok!');
+//         // FireBaseApi.authSuccess(userName);
+//         FireBaseApi.authSuccess(data);
+//         openGreetingsModal();
+//         openSignUpModal();
+//       }
+//     });
+// }
