+++ src/partials/sign-modal.html
@@ -13,7 +13,7 @@
         <label class="form-container__name" for="user-email"> Mail </label>
         <input
           type="email"
-          name="user-mail"
+          name="mail"
           class="form-container__text"
           id="user-email"
         />
@@ -28,7 +28,7 @@
         </label>
         <input
           type="password"
-          name="user-password"
+          name="password"
           class="form-container__text"
           id="user-password"
         />
@@ -87,7 +87,7 @@
         </label>
         <input
           type="email"
-          name="user-mail"
+          name="mail"
           class="form-container__text"
           id="user-email-register"
         />
@@ -102,7 +102,7 @@
         </label>
         <input
           type="password"
-          name="user-password"
+          name="password"
           class="form-container__text"
           id="user-password-register"
         />
